Refactor : 로딩컴포넌트 경로 변경
